fix(34611221): Separation Participants into components
