+++ packages/pkg.community.settings/components/Participants/Participants.tsx
@@ -1,153 +1,13 @@
-import React, { useState } from 'react';
-
-import { Badge } from '@xipkg/badge';
-import { Input } from '@xipkg/input';
-import { Button } from '@xipkg/button';
-import { useDebouncedFunction } from '@xipkg/utils';
-import { Avatar, AvatarImage, AvatarFallback } from '@xipkg/avatar';
-import { Close, CrossCircle, Plus, Search, Crown } from '@xipkg/icons';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@xipkg/dropdown';
+import React, {useState} from 'react';
+import {Input} from '@xipkg/input';
+import {Button} from '@xipkg/button';
+import {useDebouncedFunction} from '@xipkg/utils';
+import {Search} from '@xipkg/icons';
 
 // JSON со временным списком пользователей
 import usersTemplate from './usersTemplate.json';
-
-// Временные типы для Роли пользователя и для пропсов Бейджа пользователя
-type UserRoleT = {
-  name: string;
-  bgColorMain: string;
-  bgColorSecondary: string;
-};
-
-type UserBadgePropsT = UserRoleT & {
-  role: UserRoleT;
-  handleRoleDelete: (roleToDelete: UserRoleT) => void;
-};
-
-const UserBadge = ({
-  name,
-  role,
-  bgColorMain,
-  bgColorSecondary,
-  handleRoleDelete,
-}: UserBadgePropsT) => (
-  <li>
-    <Badge className={`text-xs ${bgColorSecondary}`}>
-      <span className={`mr-2 size-3 rounded-full ${bgColorMain}`} />
-      {name.slice(0, 1).toUpperCase() + name.slice(1)}
-      <Button
-        className="focus:bg-red-80 hover:bg-red-80 group ml-3 h-min rounded-full bg-transparent p-0"
-        onClick={() => handleRoleDelete(role)}
-      >
-        <CrossCircle className="group-hover:fill-gray-0 group-focus:fill-gray-0 !size-3" />
-      </Button>
-    </Badge>
-  </li>
-);
-
-// Временные типы для Пользователя и для пропсов Карточки пользователя
-type UserT = {
-  name: string;
-  nickname: string;
-  roles: UserRoleT[];
-  isOwner: boolean;
-};
-
-type UserCardPropsT = UserT & {
-  user: UserT;
-  handleUserDelete: (userToDelete: UserT) => void;
-  handleRoleAdd: (userToUpdate: UserT, roleToAdd: UserRoleT) => void;
-  handleRoleDelete: (userToUpdate: UserT, roleToDelete: UserRoleT) => void;
-};
-
-const UserCard = ({
-  user,
-  name,
-  roles,
-  isOwner,
-  nickname,
-  handleRoleAdd,
-  handleUserDelete,
-  handleRoleDelete,
-}: UserCardPropsT) => {
-  // Временный список ролей
-  const rolesTemplate = [
-    { name: 'Администратор', bgColorMain: 'bg-violet-100', bgColorSecondary: 'bg-violet-20' },
-    { name: 'Преподаватель', bgColorMain: 'bg-brand-100', bgColorSecondary: 'bg-brand-0' },
-    { name: 'Студент', bgColorMain: 'bg-green-100', bgColorSecondary: 'bg-green-0' },
-    { name: 'Гость', bgColorMain: 'bg-red-100', bgColorSecondary: 'bg-red-0' },
-    { name: '1', bgColorMain: 'bg-gray-80', bgColorSecondary: 'bg-gray-5' },
-    { name: '2', bgColorMain: 'bg-gray-80', bgColorSecondary: 'bg-gray-5' },
-  ];
-
-  return (
-    <li className="border-gray-30 flex rounded-lg border p-4 md:items-center">
-      <div className="flex flex-col gap-2 md:flex-row md:gap-8">
-        <div className="flex items-center self-start">
-          <Avatar size="s">
-            <AvatarImage
-              src="https://auth.xieffect.ru/api/users/3/avatar.webp"
-              imageProps={{
-                src: 'https://auth.xieffect.ru/api/users/3/avatar.webp',
-                alt: '@shadcn',
-              }}
-              alt="@shadcn"
-            />
-            <AvatarFallback size="s">CN</AvatarFallback>
-          </Avatar>
-          <div className="ml-2">
-            <p className="text-nowrap text-base font-medium text-gray-100">{name}</p>
-            <p className="text-gray-60 text-nowrap text-xs font-normal">{nickname}</p>
-          </div>
-          {isOwner && <Crown className="!size-3 self-start" />}
-        </div>
-        <ul className="flex flex-wrap items-center gap-x-4 gap-y-2">
-          {roles.map((role, index) => (
-            <UserBadge
-              role={role}
-              key={index}
-              name={role.name}
-              bgColorMain={role.bgColorMain}
-              bgColorSecondary={role.bgColorSecondary}
-              handleRoleDelete={() => handleRoleDelete(user, role)}
-            />
-          ))}
-          <li className="size-6">
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button className="hover:bg-brand-80 group !size-6 bg-transparent p-1">
-                  <Plus className="group-hover:fill-gray-0 group-focus:fill-gray-0 size-4" />
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent className="w-56 p-3">
-                {rolesTemplate.map((roleTemplate, index) => (
-                  <DropdownMenuItem
-                    key={index}
-                    className="hover:bg-gray-10 rounded-lg"
-                    onClick={() => handleRoleAdd(user, roleTemplate)}
-                  >
-                    <span className={`bg mr-2 size-3 rounded-full ${roleTemplate.bgColorMain}`} />
-                    <p>{roleTemplate.name}</p>
-                  </DropdownMenuItem>
-                ))}
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </li>
-        </ul>
-      </div>
-      <Button
-        className="focus:bg-red-80 hover:bg-red-80 group ml-auto h-6 w-6 bg-transparent p-1"
-        onClick={() => handleUserDelete(user)}
-      >
-        <Close className="group-hover:fill-gray-0 group-focus:fill-gray-0 size-4" />
-      </Button>
-    </li>
-  );
-};
+import {UserRoleT, UserT} from "./types";
+import {UserCard} from "./UserCard";
 
 export const Participants = () => {
   // Временное решение для рендера, удаления, изменения ролей пользователей
@@ -198,6 +58,8 @@ export const Participants = () => {
   const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
     debauncedUsersSearch(event);
   };
+
+
 
   return (
     <>
