+++ packages/pkg.community.settings/components/Participants/types.ts
@@ -1,0 +1,39 @@
+// Временные типы для Пользователя и для пропсов Карточки пользователя
+import {boolean, number, string} from "zod";
+
+export type UserT = {
+    name: string;
+    nickname: string;
+    roles: UserRoleT[];
+    isOwner: boolean;
+};
+
+export type UserCardPropsT = UserT & {
+    user: UserT;
+    handleUserDelete: (userToDelete: UserT) => void;
+    handleRoleAdd: (userToUpdate: UserT, roleToAdd: UserRoleT) => void;
+    handleRoleDelete: (userToUpdate: UserT, roleToDelete: UserRoleT) => void;
+};
+
+// Временные типы для Роли пользователя и для пропсов Бейджа пользователя
+export type UserRoleT = {
+    name: string;
+    bgColorMain: string;
+    bgColorSecondary: string;
+};
+
+export type UserBadgePropsT = UserRoleT & {
+    role: UserRoleT;
+    handleRoleDelete: (roleToDelete: UserRoleT) => void;
+};
+
+// Ответ на событие запрос списка участников
+
+export type Participant = {
+    created_at: string
+    id: number
+    is_owner: boolean
+    user_id: number
+}
+
+export type ParticipantsList = Participant[]
